KH Classify: extract Taxonomy tab + its 5 dialogs; remove shared header
Taxonomy governance (tree/duplicates/orphans/treemap/audit sub-tabs) + the
Split/Move/Merge-Into/Create-Tag/Confirm dialogs moved into
ClassifyTaxonomyTabComponent (loads its own tag/tagged-item/audit data). With
all six tabs now owning their sub-page headers, the shared
<mj-page-header-interior> + @default branch are removed; <mj-left-nav-content>
is now just a 6-case @switch. Host 3749 -> 2157 TS lines (5147 originally).
Build green.
