+++ packages/Angular/Explorer/dashboards/src/ai-dashboards.module.ts
@@ -32,6 +32,7 @@ import { ClassifyTypesTabComponent } from './AI/components/autotagging/tabs/type
 import { ClassifyTagsTabComponent } from './AI/components/autotagging/tabs/tags-tab.component';
 import { ClassifySourcesTabComponent } from './AI/components/autotagging/tabs/sources-tab.component';
 import { ClassifyPipelineTabComponent } from './AI/components/autotagging/tabs/pipeline-tab.component';
+import { ClassifyTaxonomyTabComponent } from './AI/components/autotagging/tabs/taxonomy-tab.component';
 import { TagsResourceComponent, LoadTagsResource } from './AI/components/tags/tags-resource.component';
 import { DuplicateDetectionResourceComponent } from './AI/components/duplicates/duplicate-detection-resource.component';
 import { VectorManagementResourceComponent } from './AI/components/vectors/vector-management-resource.component';
@@ -94,6 +95,7 @@ import { MJWordCloudComponent } from '@memberjunction/ng-word-cloud';
     ClassifyTagsTabComponent,
     ClassifySourcesTabComponent,
     ClassifyPipelineTabComponent,
+    ClassifyTaxonomyTabComponent,
     TagsResourceComponent,
     DuplicateDetectionResourceComponent,
     VectorManagementResourceComponent,
